+++ src/main/groovy/io/xh/hoist/ldap/LdapGroup.groovy
@@ -1,19 +1,21 @@
 package io.xh.hoist.ldap
 
-import javax.management.Attribute
+import groovy.transform.CompileStatic
+import org.apache.directory.api.ldap.model.entry.Attribute
 
+@CompileStatic
 class LdapGroup extends LdapObject {
 
-    String cn
+    /** DNs of all group members (yes, the property name looks singular, but is the collection) */
     List<String> member
 
-    static LdapGroup create(Collection<Attribute> atts) {
+    static LdapGroup create(Collection<Attribute> attributes) {
         def ret = new LdapGroup()
-        ret.populate(atts)
+        ret.populate(attributes)
         ret
     }
 
     static List<String> getKeys() {
-        LdapObject.keys + ['cn', 'member']
+        LdapObject.keys + ['member']
     }
 }