+++ src/main/groovy/io/xh/hoist/ldap/LdapPerson.groovy
@@ -1,21 +1,16 @@
 package io.xh.hoist.ldap
 
-import javax.management.Attribute
+import groovy.transform.CompileStatic
+import org.apache.directory.api.ldap.model.entry.Attribute
 
-/**
- * Minimal set of attributes returned by a search for objectCategory=Person in Microsoft's AD LDAP implementation.
- * If you need more attributes, you can extend this class and add them.
- *
- * Available attributes can be found here:
- * https://learn.microsoft.com/en-us/archive/technet-wiki/12037.active-directory-get-aduser-default-and-extended-properties
- */
+@CompileStatic
 class LdapPerson extends LdapObject {
 
-    String name
-    String displayname
+    /** First name of the person */
     String givenname
+
+    /** Last (sur)name of the person */
     String sn
-    String mail
 
     static LdapPerson create(Collection<Attribute> atts) {
         def ret = new LdapPerson()
@@ -24,6 +19,6 @@ class LdapPerson extends LdapObject {
     }
 
     static List<String> getKeys() {
-        LdapObject.keys + ['name', 'displayname', 'givenname', 'sn', 'mail']
+        LdapObject.keys + ['givenname', 'sn']
     }
 }